src/api/schema.py
import strawberry
from typing import List, Optional
from .models import GenomicQuery, AnalysisResult

@strawberry.type
class Query:
    @strawberry.field
    def analyze_sequence(self, query: str) -> AnalysisResult:
        # Implement genomic analysis
        pass

    @strawberry.field
    def query_status(self, query_id: str) -> str:
        # Implement status check
        pass

schema = strawberry.Schema(query=Query) 